Add overflow handling for timestamps and ports
If the input data is too large for the target type log an error and set
the target data to 0.

The added checking should remove the need for te #nosec G115 comment but
is seems the nested dnstap struct is causing issues for gosec, see the
linked issue comment for details.