Update new_qname message to latest version
Drop use of automatically generated struct. The go-jsonschema tool seems
to not understand the "const" keyword at this time, so with the latest
version of the spec the "type" field was converted from a string to
interface{} which just complicates everything.

Lets just do things by hand for now.

Also make golangci-lint happy (converting config to v2) and then dealing
with capitalisation of things like "JSON" and "ID".

Fixes #42