Make parquet tests always run
Instead of only running parquet writing tests if supplying a flag, make
the tests always run but write to an internal buffer instead, and only
writing out actual files if the flag (-write-parquet) is supplied.

This way we should catch if e.g. the sessionData struct and
sessionDataSchema goes out of sync without having to always write out
files.