getClientCertficate(): return error if cert is nil
While this should not happen as we call setCert at startup lets make the
function behave properly if it is. This fulfills the expectation from
the docs for tls.GetClientCertificate:
```
If GetClientCertificate returns an error, the handshake will be
aborted and that error will be returned. Otherwise
GetClientCertificate must return a non-nil Certificate.
```