Add test for HLL explicit threshold
Have a test to verify that we only include HLL data in histograms once
we pass the HLL explicit threshold.

While adding the test it became clear that the existing
writeHistogramParquet() function was not really written to be easily
testable since it handled writing out files which is cumbersome to deal
with from test code. Now the code has been rearranged so
writeHistogramParquet() only receives a io.Writer and file handling is
done in an outer createHistogramFile() function instead.

This means the test code can now just supply it with a bytes.Buffer and
inspect the parquet data in there.

Since writeSessionParquet() is bascially a copy of
writeHistogramParquet() do the same change there introducing
createSessionFile() as its file handling wrapper.

Introduce newHistogramData() so the main code and tests can create such
structs in the same way.

Finally make a const out of labelLimit (now defaultLabelLimit) since it was
needed in the test.

Fixes #41