Update version handling
Up until this point we have created RPM packages with versions like
"0.0.0^20250922.51f14e3". The edm binary in these packages have not
had the main.version variable set since this has never been done with
"make build".

Make it so the version logged by edm is either the full sha revision, or
a tag like "v0.0.1", while the rpm package version is set either like
a snapshot ("0.0.1^20250922.51f14e3") or just the tag version
("0.0.1") if the specific revision we are building from is tagged.

Move the logic into a shell script since it became cumbersome to do more
involved stuff directly in the Makefile and make the go build step
set -ldflags based on if the VERSION file exists or not.

Discussed with @zluudg