Converge on "dnstapir-edm" naming for the binary
Before this we have slowly diverged where the repo calls the tool "edm"
while the rpm packaging calls it "tapir-edm". After discussions we have
decided it is better if the binary is called "dnstapir-edm" to make it
stand out better especially when being installed from a package.

Move things around so the "main" package lives under cmd/dnstapir-edm and
update surrounding code and package specs to reflect the new tool name.